Revert "gd: Remove CounterMetrics as dependency of StorageModule"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/11ae73b5-7a51-4f9d-9671-62943aaa92b7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 386842159
Bug: 386843345
Change-Id: I4e6d1eaa3d181f3e49378d694f04991143097600
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com