Revert "hcidoc: fix the number of ACL connections"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4c30226d-4403-4a9d-8f71-82dabc3ddbdc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8438598598c160b7f27fc8b8d648a342b8ea5685
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com