Revert "floss: Parse manufacturer data as little endian"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e6b78d83-cdd9-4c8d-bfab-697cc298cb2f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8cf14c5fda424e4fc4df37a02f9c8107c85be859
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com