Revert "Add a flag to fix ATT_EXECUTE_WRITE_REQ"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9dd6021b-cf2a-4395-ac33-55400ef2307f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I443a72e50953c997440c1d34d614fcf5d150dc0b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com