Revert "CS: Do not omit any CS subevents when sending CS enable"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9dd6021b-cf2a-4395-ac33-55400ef2307f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9f377a8a08ced39183e7e578cf48680d47378490
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com