Revert "Send ATT_EXECUTE_WRITE_RSP when no pending writes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9dd6021b-cf2a-4395-ac33-55400ef2307f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie188a2c0b06d33625eb1f798d60ae059fdd61787
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com