Revert "hcidoc: fix the number of ACL connections"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/84f9c821-662a-4b13-99a5-0a31f88a4825). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I22277d61fcc9d2e335b8bc17014cf2bf38135635
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com