Revert "floss: save address privacy setting to backup file for r..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e42683a5-d341-4dfc-adee-51335c4a2cdc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 387001264
Bug: 387001264
Change-Id: Ic5d2a2f51bf96c2657beaea72485cd25d842f12d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com