Revert "Add new metadata APIs get/set for is_microphone_for_call..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e0a775fd-a43a-4883-bec5-a4d7c0dc3eb6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I068576550562ddedf1050c76e100290f45d70100
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com