Revert "Fix some VisibleForTesting usage problem in Metadata"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e0a775fd-a43a-4883-bec5-a4d7c0dc3eb6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic396a2065bf56c5346b69e2b62ebace4647898f9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com