Revert "Flag to create device metadata only when device is bonded"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ab1c591a-29dd-442b-b039-ed21d19550c4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I423736be69cb87fd65436124efb99a913b4fbaa1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com