Revert "Flag to create device metadata only when device is bonded"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8d995621-e88c-484d-a20f-3bce951e800d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I09fa51626875c3f50edd392967d4d1795de5c612
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com