bta: csis_client: Remove device when bonding failed
This fixes the issue where the device being in bonding state
is disconnected abruptly, leaving the bonding information native.
Such device could not be automatically paired then, because
in CSIS native the bonding process was still in progress.
This adds new bondingFailed() method to be called from upper layer
to notify the native layer about the failed bonding procedure,
so that the native could clean up the context.

Bug: 357472821
Flag: Exempt; regression covered with unit tests added
Test: atest bluetooth_csis_test
Change-Id: I77f179585cce4de9ccf0be32dc476038f35bbe29