Revert "Replace the flag le_enc_on_reconnection with the flag le..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2a54ad72-a4cf-4ce2-a99d-a468ee94e5b0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 389537880
Change-Id: Id9edca67bc037c162d4504231db6b4f9615b3525
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com