Bass: Always execute suspend procedure if external broadcast is managed
If broadcast was suspended (STATUS_LOCAL_STREAM_REQUESTED) and resuming
(STATUS_LOCAL_STREAM_SUSPENDED) was not finished before new suspend
came, then suspend was not executed.
Because paused broadcast sinks data
are cleared at the beginning of the resume, then we lost this
information.

With this CL, suspend procedure will always be executed if the primary
device has managed external broadcast, so sinks will be properly cached.

Bug: 378615357
Bug: 378661060
Test: atest BassClientServiceTest
Change-Id: Ief1875712474cbae879c9e8445724bdcc03dad30