Revert "Extend and randomize BTA_DM_SWITCH_DELAY_TIMER_MS"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4ecf4c3d-db8d-4625-9b5b-3334844c8e71). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If345a3fa142a48a22cd9f5905915d7bd57f49172
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com