Revert "Add flag extend_and_randomize_role_switch_delay"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/54c48550-63c2-4975-bfe2-24ede79a1aa1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I479c02c79c3e0d92601b071d783d6ca73f8b6fbb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com