Revert "Print the attributionSource used for permission"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a6c05921-a4b4-4454-b38b-4ecfb51c0db6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iad326b4d4f463c4e6125725558eeebc89bde3fbf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com