Revert "Moved isWatch() to Utils"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a6c05921-a4b4-4454-b38b-4ecfb51c0db6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1f491dfc0866925df3991fdaa684deec38fcaa4c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com