[Upstream] Prevent AVRCP play event when voice active.
Add a check for voice communication state before sending a
play event to AudioManager as some device may send a play
event when SCO is disconnected but the call hasn't ended.

This is a BT specific use-case so we can fix it in the BT
stack instead of the Audio framework.

Also did a small cleanup and refined the logging to reflect
the recent changes to sendMediaKeyEvent.

Bug: 395534976
Test: atest BluetoothInstrumentationTests
Flag: Exempt Trivial
Change-Id: I410b097aa385c61bf3fbdf40c42e9ed1e11b51a1