Revert "a2dp_test: add test for avdt collision as initiator"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a2ab16dc-2c20-495f-b84d-65f36697b176). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7ca1dda73cce2c12ff50098d40820f43def8d226
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com