LeAudioService: Fix setting inband mode when switching LeAudio HS
Fix setting inBandMode in TBS status flag when switching Active device
between two LeAudio Headsets.

Without the patch, group which became Inactive was not updated with a
new inBandMode, which is necessary for the device to have in order to
generate internal ringtone on Incoming call, as inband tone will be send
only to Active device.

Bug: 385083184
Flag: Exempt, day-to-day fix. Regression test with unit tests
Test: atest LeAudioServiceTest

Change-Id: I44281abf5962ceb91ddac6cd5dd1af227c718c99