Revert "Print the attributionSource used for permission"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ca73bed3-a66c-44e5-95eb-9d68b2cbaa09). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I844c6e398637ca5e429e53c75987dd2a2c6b6b86
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com