Revert "Add new flag for change default trackable adv number"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/32505431-1f35-4522-91d4-2b8896caefc6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0225c0afdc40aee4968294b87fb92ba19647fdb8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com