Bass: Refactor handling select source requests
Collect the variables that depend on syncHandle and modify them
together where possible. Add synchronization to ensure they are
updated safely. This prevents unexpected behavior from concurrent
modifications. Allow the creation of isAddedToSelectSourceRequest
since they are modified simultaneously. In the future, this will
simplify refactoring to create a single object to store them together.

Ignore synced broadcast in onSyncEstablished if there is no pending
synchronization. There is no possibility to unsync it as we do not
have mPeriodicAdvCallbacksMap for it.

Bug: 389722679
Flag: EXEMPT, refactor covered by unit tests
Test: atest BassClientServiceTest
Change-Id: Iddf2dfdd880b6c9594e6d8e627d9ecabe610cc9e