Revert "Clean up flag scan_record_manufacturer_data_merge"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4f60d3ca-3003-4f61-9157-fc29f5e30d2c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I581525d43e63dd25214cc8001dd9e9c3177150c9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com