Change default number of trackable advertisements for onFound filter
The current default number of trackable advertisements set for each
onFound filter is 'MATCH_NUM_MAX_ADVERTISEMENT', which is 'Max hardware allows / 2'. This is exhausting
resources and causing errors. Therefore, changing the default to
'MATCH_NUM_FEW_ADVERTISEMENT', and also changing the value of
'MATCH_NUM_MAX_ADVERTISEMENT' to 'Max / 4'.

Bug: 386727721
Bug: 389568695
Test: atest ScanManagerTest, ScanSettingsTest
Change-Id: I39cb19fd6e299c9b74c441fc3da35f232cc8ec74