Close all RFCOMM sockets if BREDR is disabled but BLE ON mode
This is a fix issue when bredr is off and ble is on, but rfcomm socket
is still listening for RFCOMM connections from the devices.

Call the new API after BREDR is off but Staying in BLE_ON mode in
BluetoothManagerService.java.

Change-Id: I7050a82d812a117c4c2ca806c9a7eacec93dc772
Flag: com.android.bluetooth.flags.rfcomm_socket_closed_by_bredr_disabled
Test: m com.android.btservices
Bug: 387198958