Revert "Increase LeScanningTest TIMEOUT_SCANNING_MS to 3000"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/058e2b1e-313e-4ca6-9fc9-31f0ec4bac08). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8ad284eea0abfa8af5b2b50ac5d7b6ba931acaf9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com