Revert "leaudio: Inactivate device on autonomous stream release"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eb18ef99-b9df-471f-ace2-fc38dec40028). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I26d472c7a6b75667411d5e55cc7d5d2f29991618
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com