le_periodic_sync_manager: Terminate sync if there is no such sync index
It could happen if we first stopSyncs which are established, and after
that cancel pending sync. It looks like
LePeriodicAdvertisingCreateSyncCancel somehow override
LePeriodicAdvertisingTerminateSync. We should stopSync again as we are
not interested in it anymore if there is no such sync index.

Bug: 392572928
Test: atest PeriodicSyncManagerTest
Flag: EXEMPT, day-to-day fix covered by unit tests
Change-Id: I380094f0e8f826d9503b439f19f4c94807f85c49