Revert "Reject SMP pairing SMP_BUSY error while host is in pairi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/14281937-b02a-4cec-b18f-25b5fec6e197). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1210ebc7a7a36e42bf1cf6ba668e518c8b0349e6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com