Revert "Unregister GATT clients when disconnected"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b04e944c-fadb-47f9-b6a9-2c03ff4ba45f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia750515ff7a54e6d3597cc9fbb06990c1794c1a1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com