Revert "BluetoothGatt: Fix duplicate GATT services being added"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b04e944c-fadb-47f9-b6a9-2c03ff4ba45f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I140dceaa34ae484b6fc7baefabd7510fbbf7bff6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com