Revert "flags: adm_remove_handling_wired"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d9fdb8ae-11f9-482d-81b1-b8f81f12a8eb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2115e6380c4326be19b396bc30a28cefb0ccb591
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com