Revert "Increase LeScanningTest TIMEOUT_SCANNING_MS to 3000"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d9fdb8ae-11f9-482d-81b1-b8f81f12a8eb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I91fa52a2e67f8fc9ba5b670e8b1ae888d0e9e160
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com