Add interop for SetAddressedPlayer changed notification
Some carkits don't handle receiving a changed notification
properly and will loop in setAddressedPlayer. We need to
have an interop for those.

Note: this change is not directly flagged but should be
considered as such because this part of the code isn't
reachable without a flag enabled.

Bug: 393059901
Bug: 379039786
Bug: 381719439
Bug: 352279986
Flag: com.android.bluetooth.flags.set_addressed_player
Test: atest BluetoothInstrumentationTests
Change-Id: I445d3486bede96f73c7b1bb3fb7d62915dfa3d93