Revert "leaudio: Inactivate device on autonomous stream release"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/54284c80-4919-4cc3-8887-1ba1f4695f97). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic596efdfac38d49e530cb0db4d58d151bdcbc40d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com