Revert "system/stack/btm: Enforce -Wmissing-prototypes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b7e2086f-4eba-4901-b2a7-80b8e4408806). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1bfa439f2dd2f6866b75b3e8b8296fe8e4a4e2b2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com