Revert "Add flag: key_missing_ble_peripheral"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/98fcf294-0cf6-4500-9e45-f1705d775621). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iabeea1ebfa4d9ca59a4f2ee319ef9043f929d30b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com