Revert "Don't call cleanup in GattService#stop."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/25f4ae8e-9bdc-4cfd-96ab-bded85dc0c11). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie96245897cd611c7a229847bc0fa8877f6d61519
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com