Revert "avdt_accept_open_timeout_ms isn't a flag but a sysprop"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b4e2744f-46c9-4eaa-8d2e-49a63045ffb8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie5d66dc14434ad8513fab1efd626bba7cf0481a4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com