Bass: Remove sink broadcast metadata in source switch during suspend
Follow up on "Add support for multiple sources per sink device"

Sink metadata must be removed as it is no longer valid.
If not removed, the sink may attempt to synchronize with it
during resume (for another broadcast) or reconnection.

Even if there is a room for broadcast, it could happen that all
broadcasts were suspended via removing source. In that case, we have
to found such broadcast and remove it from metadata.

checkAndStopBigMonitoring will not remove HOST_INTENTIONAL pause type
if there are any paused sinks to give possibility of check it.

Bug: 385389212
Flag: EXEMPT, day-to-day bugfix covered by unit tests
Test: atest BassClientServiceTest
Change-Id: If4a4cb584d1d0b990d68f5a5c912ff68e484257a