Bass: Remove sink broadcast metadata in sink disconnection
Follow up on "Add support for multiple sources per sink device"

During disconnection, sink metadata cannot be removed directly,
as it is used for synchronization after the device reconnects.

If the entire group is unsynced, it must be removed since
resynchronization is not expected. To achieve this, the whole
group should be checked, and if all devices are unsynced,
sink metadata should be removed from the group, including
previously disconnected sinks whose metadata was retained.

Bug: 385389212
Flag: EXEMPT, day-to-day bugfix covered by unit tests
Test: atest BassClientServiceTest
Change-Id: Ifaf765c401bb5f5aaf3cf73e7b15fad59cabf839