Revert "Flag to handle LE encryption refresh failure"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7c09e011-50c8-495f-a625-6ade070c3a48). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id948e3a8557e93e6e40a1203bed1525eff93884a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com