Revert "Remove bta_dm_discover_both flag"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7214f07e-5997-4876-a5d9-a4ed65e6411f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibd4f0ee2b65eccecb145d1041cd731135315c14b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com