Revert "BluetoothService: Respect config_hsumBootStrategy"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/660e66e4-3a73-4808-9a04-2316830dba68). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9f7698611a616b796df135cb2ad2fd801f188b0c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com