Revert "Add flag: key_missing_ble_peripheral"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5ad3a2f4-98ff-4d67-a00d-b01aa5fe1acc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id40759c8ef3ebffb88845910acd830df6b125986
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com