Revert "Add flag: key_missing_ble_peripheral"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c48cafb5-3ee9-4941-b122-13378469348c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I421d0f2acf8d82347ad5d63430f2e4394c871e5d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com