Skip the all-zero IRK in the RPA resolution process
[Root Cause]
The BT stack uses an all-zero IRK to resolve RPA.
This will cause the upper layer to mistakenly replace the information in the Advertising Report, such as the device name, etc.

[Solution]
According to the Bluetooth Core SPEC, an all-zero IRK indicates that the device does not have an RPA.
It is recommended to skip the all-zero IRK in this process of the BT stack.
1. If the IRK is all zeros, do not mark the PID key type as available.
2. Only when both the PID key and LID key are invalid, do not add the address to the Resolving list.

Bug: 388222528
Test: atest net_test_btif pass
Change-Id: I2031e695a67bb05370eaf2ce7d6718cac04507c4