Floss: Separate cleanup procedure on disabling adapter and abort
When the controller is in a bad state, ex. HCI command timeout or
on Hardware error event, Floss should teardown the process without
sending other commands to controller to prevent further error. This is a
different behavior from the normal disabling adapter, where we do want
to set discoverable or connectable before adapter shutdown.

This CL separates the above 2 cases by handling in different signals.
This also increase the log verbose in the clean up state to make the
debugging easier.

This CL also separates the API shutdown and adapter shutdown as they are
doing different things.

Bug: 393662321
Tag: #floss
Test: mmm packages/modules/Bluetooth
Flag: EXEMPT, Floss-only changes

Change-Id: I05710800e57680d1c77b09ea70d5ffbb032a34a5