Add default player to AVRCP player list
A previous patch got rid of the "Bluetooth player" that
wrapped over all Media Players. By doing so, the playerID
of the Media Players was brodcasted to the remote device
but started at ID 1. This seem to be an issue for some
devices and we need to make them start at 0.

Given that the Bluetooth specification also mentions that
there should always be a player on target, AVRCP has to add
a "default" player that should be available even if no
Media Player is available on target.

This change adds back the fake Media Player with ID 0,
making it non-browseable, and keeping the real list of
Media Players "outside" of it.

The response to a GetFolderItems - MediaPlayerList will now
be: "0 - Bluetooth Player", "1 - Real Media Player"... and
so on.

Note that as the "Bluetooth player" is not browseable, it
will not appear on the Brwse UI on devices and should be
totally transparent to users.

Bug: 394035701
Bug: 338156685
Test: atest BluetoothInstrumentationTests
Flag: com.android.bluetooth.flags.browsing_refactor
Change-Id: Ie6ea82f5386635e1095d0dac4df54c3c180d3d31