Revert "system/stack/btm: Enforce -Wmissing-prototypes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3ac5ca3c-9a6f-46a0-8960-c83284afc879). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I351796c90b466a6882bb94f1d021658eee22d7ee
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com