Revert "BluetoothService: Respect config_hsumBootStrategy"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9b01daf8-7586-4cf0-aefb-f4dae887adaf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ida50f55f5ab81c64231c74db4f31e30254a6085f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com