Revert "Don't call cleanup in GattService#stop."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9b01daf8-7586-4cf0-aefb-f4dae887adaf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I538b854df6d3de6926cc2b186adc74e67f994112
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com