commit corpus for rfcomm-fuzzer validation test
These files form a set of "known good" inputs to the fuzzer
and are not expected to find any new memory safety issues.

The fuzzer should run on all of these files without failing,
any failure is a regression that will be caught in presubmit.

Bug: 395162086
Flag: EXEMPT resource only update
Test: m com.google.android.btservices
Change-Id: I29e38d451e3e5e19abee2fa613492ed0a4ef01e7