Revert "CS: add test for RAS server"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eabc8526-d747-4290-8dd4-996b400b7f5c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib57754699b07f48818422f533767d10a529a47f4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com