Revert "Fix to update inband ringtone status for BT connection d..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f08d120f-ca0f-4150-a581-83afaac03098). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ieea30ac9513b0ca6acbe94583a0fd5031a9dc4f4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com