Revert "bta_av: avoid trying to get codecs before the stream starts"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/25f40b53-2de8-418a-8970-0d05e2bde010). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If5fea93bbb6f52de4d9ae599c07ae0f631b9afb0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com