VC: Allow to set volume if different from current or started pending OP
Follow up on "Allow to set the same volume during already pending
operation"

Always remove any pending (but not started) volume operations when a
new request arrives. Before proceeding, check if the current device
volume or the volume of an ongoing operation differs from the requested
value.

When a user requests a new absolute volume and there is an unexecuted
pending request in the native stack (limited to one per device), the
pending request should be replaced to reduce latency and avoid
unnecessary traffic.
Furthermore, if the requested value matches the volume on the remote
device or the currently executing request, it can be skipped to further
minimize unnecessary traffic.

Bug: 391826441
Bug: 386758535
Test: atest --host bluetooth_vc_test
Change-Id: Ia42d5fce4738020e15417c3b2fb6588e091b1458