VC: Allow to mute/unmute if different from current or started pending OP
Always remove any pending (but not started) mute/unmute operations when
a new request arrives. Before proceeding, check if the current device
mute state or the state of an ongoing operation differs from the
requested value.

If a new mute/unmute request is received while a pending request exists
in the native stack (limited to one per device), the pending request
should be replaced to reduce latency and unnecessary traffic.
Furthermore, if the requested mute state matches the remote device or
the currently executing request, it can be skipped to further minimize
unnecessary traffic.

Bug: 382213180
Bug: 386758535
Test: atest --host bluetooth_vc_test
Change-Id: I5c2197bf7b70320df6e2e0a0d20c57a34eca0c08