system/rust: Remove static globals from the Rust GATT server
The use of global variables is prone to error and was arguably
the root cause of http://b/356462170.  This issue is compounded
by the Rust module being a module, which meant the dependencies
were tricky to keep track of.

This CL refactors the initialisation and destruction of the private Rust
GATT server.

TODO --- add more here.

Bug: 356462170
Test: m com.android.btservices
Flag: EXEMPT refactor only
Change-Id: I46b5d9858646e65e435d3a091683d866a003969d