Revert "Split Bluetooth into an android library"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d0dc079e-93d7-4c29-9f9a-bc26f659cad2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 395811038
Change-Id: I3ea837405c29c6972abf2668d7e0de4f1f8bc065
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com