Bass: Monitor OOR broadcast during scanning in onScanResult
When the sink is unintentionally paused and the broadcaster is out of
range, monitor it in onScanResult during scanning instead of attempting
to sync in a loop. If the scanner is disabled, continue attempting to
sync in a loop. That allow to sync to another broadcasters.

Bug: 390189678
Flag: EXEMPT, day-to-day bugfix covered by unit tests
Test: atest BassClientServiceTest
Change-Id: Id42081bedadb76706164f7abad91d16f8d1a9c85