Create a rule wrapper around mockito
Today, the test are directly using the mockito rule and that works fine.
But in order to switch to mockito-target-extended, we need to call the
Mockito.framework().clearInlineMocks() at the end of each test.

We don't want to add it manually, so instead we are creating a wrapper
to the mockito rule and we are cleaning in one place, to activate for
each test.

This cl DOES NOT activate the mockito-target-extended yet, but this
change is an automated refactor to facilitate switching to
mockito-target-extended.

See https://r.android.com/3468588 and associated changes to a completion
of this task

Bug: 191783947
Test: m BluetoothInstrumentationTests
Flag: TEST_ONLY
Change-Id: I2d8bb10dbade1ed64a44011517d69d0efc336df9