Revert "bta_av: avoid trying to get codecs before the stream starts"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f97fab30-874f-4a62-90ce-ae71e04f9aca). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0180abb9bc8bb4216db99d5d1567612fbac4d9cf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com