Revert "le_audio: Do not configure device that has not contexts ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/47ba429e-9aa0-4981-ab60-b009d606711d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I55355aa92ac09b89f727e8fc8aba1187f3c7d3ed
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com