Revert "MapClientService: merge `stop()` into `cleanup()`"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e973518f-5872-484c-bfab-9c9751835086). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 396783961
Change-Id: Ifeb252e3ee5f7c52f4d18d7018f9b9041b731719
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com