Revert "MapClientService: merge `stop()` into `cleanup()`"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eba11b74-5ca7-483a-93aa-27134f6b3891). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 396784418
Change-Id: I85166514104c3cc519dd68d3f1aa86ad600a5a22
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com