Revert "MapClientService: merge `stop()` into `cleanup()`"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f556d0ba-147f-4ab9-8e22-3fd940e39092). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 396788388
Change-Id: Ie824ebd963ba8785ea5360f74220ff9d6605850f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com