Revert "BluetoothGatt: Use CopyOnWriteArrayList for thread safety"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/68770b29-4496-4b9f-a9d4-febada340686). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I397ab7c79e8ed024fefd27c7a242b1bdf75db4a3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com