Revert "le_audio: Do not configure device that has not contexts ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8b2f331d-c899-4aeb-bbe9-3aaec6059cdf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie0bebb48b4b94ffff1be59b4aac12c966e60b930
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com