Revert "Change instance field names from "util" to "mUtil""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6fef27fb-a39d-4968-8bab-7ae820632b67). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib15b51b8396a6f71837dc90d37d7b1442a2b4afc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com