Revert "[bass] Null check when setRssi to metadata"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6fef27fb-a39d-4968-8bab-7ae820632b67). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2212298f8b0c5ff04e2650252b765f40a92a12c8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com