Revert "GTBS: Fix for long read"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/63a887e5-9e00-460b-8ee7-e6263e4e34c8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I848770e27058f9b6068e48ebe57a15fba7e5a426
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com