Native bond status interfaces should consider persistency as well
IsDeviceBonded() does not check if the pairing was persistent or not. Also the bond_type parameter is relevant only for BR/EDR.
Also BTM offers two sets of bond status interfaces. IsLinkKeyKnown() can be used for either transport but does not consider the temporary pairing. IsDeviceBonded() just checks if the device is bonded over any transport.
Both of these interfaces are now consolidated into a single interface which allows checking transport specific bond status or bond status irrespective of transport.

Change-Id: Icef05f847de04b7fc06e741313da88a48d2efe1c