Revert "Add flag empty_names_are_invalid"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/da37337b-602a-443d-910b-78609002e6cf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9c5dc81e3545e7cf4baa5087c8852c9c3ddbaf67
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com