Revert "Add @VirtualOnly annotation for startBleScan_withService..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0479db7d-d551-49de-bee0-1f0da5bd414d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I27570fdfedd8f2c2f9c05991b5c011b470595dfd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com