Bass: Prevent adding already synced broadcast in addSelectSourceRequest
This could occur when an unintentional paused sink is triggered by a
music pause while the remote unsyncs from BIS and PA. Subsequently,
addSelectSourceRequest is called on every onScanResult, even if the
source is already synced.

It was checked in handleSelectSourceRequest, but there was still
unnecessary adding and retrieval from the queue.

Check directly for the existence of a syncHandle in
getActiveSyncedSources instead of creating a new list if it's not
modified or used multiple times

Bug: 390189678
Flag: EXEMPT, day-to-day bugfix covered by existing unit tests
Test: atest BassClientServiceTest
Change-Id: I8056a9ae048e52a87b3bd15e3ef60799d7edcea4