Track and set the pairing persistency for all BR/EDR pairing
Track whether the pairing type is persistent or not for all pairing. Ensure that pairing type is correctly applied to the device after BR/EDR pairing is complete.
Pairing persistency is applicable only for BR/EDR transport at present. It must not be set for LE transport pairing.

Test: mmm packages/modules/Bluetooth
Flag: EXEMPT bugfix
Bug: 396433944

Change-Id: Icf99c12d8d78f01a32c881ba6a2dc8ad6b6978da