Revert "le_audio: Do not configure device that has not contexts ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3043bb00-fd85-4374-94a6-9f077085e1fa). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I685d6270fb1067104dd24d18b45b8b50d3123b42
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com