Revert "btif: find all services when checking profile support"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9b9005a2-4d63-4936-a870-e0c93276c82c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib8a526006e156c6e24ca22c3fe5d576e627ee378
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com