Revert "GMCS: Fix for long read"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6e979f40-6cc9-41bb-ab0e-a30f0804a151). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I151475b8c4edbd333d6bc0b5372be872f37c4da0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com