Revert "avdt: add timer for monitoring delay report as initiator"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4f6d8c23-9287-4160-884e-66ef47c4e811). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 398000890
Bug: 398006073
Change-Id: Iac1982792b63d8d328014e487f882af66e8b0ed1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com