Revert^2 "Run unit test with sources"
This reverts commit 2cec0df5e5475e5e929177de7f7964fb86433329.

Reason for revert: reland

Run unit test with sources

Also, switch to mockito-target-extended.

For years, Bluetooth java test were trying to run against the on device
code. The prebuilt process and now the mainline process can lead to
failure in case the Bluetooth code is out-of-sync from the branch being
tested on.

This CL fix this. Now the test will recompile the source code and when
run, will guarantee that the code being tested is from the same branch
and release as the test of the code.

The side effect is that some test actually require the permission given
by the real bluetooth (threw instrumentation and shared user ID). I am
moving those into the instrumentation folder. It would be a next step to
rework these tests into a more flexible test.

Another next step will be to transition the unit test to bivalent
robolectric test, to allow host testing of the java test.

Another benefit of such change, is that the testing on Release will be
able to run properly the test with @EnabledFlags while keeping an
optimized Bluetooth app

Bug: 345013072 
Bug: 393810023 
Change-Id: I01491d7f51e96a35bd4e7c3cf661c6bdcddbc459
Fix: 393810023 
Fix: 191783947 
Flag: TEST_ONLY
Test: atest BluetoothInstrumentationTests
Test: atest BluetoothUnitTestsNonUpdatable