Fix transport selection for BT pairing with dual mode devices
When createBond API is called with Auto Transport, BR/EDR transport is selected for pairing with the dual mode devices using the public address. This can result in pairing attempt failure if the remote device is not connectable/pairable over BR/EDR.
This change chooses transport for BT pairing based on the transport seen
during the discovery of the remote device.

Change-Id: I0bb41006049edae1144d196c2911c7f895452014
Test: atest PairingTest
Flag: EXEMPT bugfix
Bug: 388074343
Bug: 398329326